requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
backports.zoneinfo==0.2.1
boto3==1.24.37
botocore==1.27.37
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.6
django-cors-headers==3.13.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tk==0.1.0
Twisted==22.4.0
<<<<<<< HEAD
# twisted-iocpsupport==1.0.2
=======

>>>>>>> 4d986cd74804b2c546c04d4d86ed40095010719b
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
wrapt==1.14.1
zope.interface==5.4.0
